# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Eligo Backend API
  version: 1.7.0
  description: |
    Eligo 后端接口的唯一机器可读契约。
    M1 个人资料契约采用“同一账号叠加能力”模型：普通参与和个人发布不拆分账号身份，
    用户与后端预置企业建立有效 OWNER 关系后，再叠加企业操作能力。
    当前不提供企业申请、认证、审核或身份切换接口。
    新增或修改接口时，后端开发者必须在同一个提交中更新本文件。
    除文件内容下载接口外，所有 JSON 响应都使用统一 Result 包装。
  x-generated-from-code-date: 2026-08-22
  x-contract-updated-date: 2026-08-22
  x-m1-contract-status: REVIEW
  x-m2-contract-status: IMPLEMENTED
  x-m3-contract-status: IMPLEMENTED
  x-m4-contract-status: IMPLEMENTED
  x-m4.1-contract-status: IMPLEMENTED
servers:
  - url: http://127.0.0.1:8080
    description: 本机 Docker / 微信开发者工具
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/agreements/current:
    get:
      tags: [Agreement]
      operationId: listCurrentAgreements
      summary: 查询当前生效协议
      description: 匿名可访问；带有效令牌时 agreed 返回当前用户的同意状态。
      security:
        - {}
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AgreementType"
      responses:
        "200":
          description: 当前协议列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgreementItemsResult"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/agreements/{agreementId}:
    parameters:
      - $ref: "#/components/parameters/AgreementId"
    get:
      tags: [Agreement]
      operationId: getAgreement
      summary: 查询协议正文
      description: 匿名可查看当前生效协议；登录用户还可查看自己曾同意的退休版本。
      security:
        - {}
        - bearerAuth: []
      responses:
        "200":
          description: 协议详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgreementDetailResult"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/agreements/{agreementId}/consents:
    parameters:
      - $ref: "#/components/parameters/AgreementId"
    post:
      tags: [Agreement]
      operationId: consentAgreement
      summary: 同意一个当前生效协议版本
      description: 首次同意返回 201；重复提交幂等返回 200。
      responses:
        "200":
          description: 已存在同意记录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgreementConsentResult"
        "201":
          description: 已创建同意记录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgreementConsentResult"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AgreementId:
      name: agreementId
      in: path
      required: true
      description: 十进制大编号；JavaScript 客户端必须按字符串保存。
      schema:
        $ref: "#/components/schemas/Id"
...
  responses:
...
    Error:
      description: |
        统一错误响应。常见业务码：
        10000-10999 通用错误；11001-11005 登录会话；
        11101-11103 手机与设备；11201-11203 资料；
        11301-11302 协议；11401-11404 文件；11501-11503 账号数据；
        11601 幂等键冲突；11602 活动状态冲突；11603 乐观锁版本冲突；11604 活动容量冲突；
        11605 幂等请求对应的活动草稿已删除。
        11701 报名尚未开始；11702 报名已经截止；11703 活动名额已满；
        11704 活动发起者不能报名自己的活动；11705 活动已经开始；
        11706 活动已取消；11707 活动已结束；11708 当前参与记录不可由发起者移除；
        11709 已被活动发起者移除，不能再次报名；11710 不符合活动报名性别限制。
        11801 不能关注自己。
        11901 动态创建幂等键冲突；11902 动态状态冲突；11903 动态版本冲突；
        11904 幂等请求对应的动态草稿已删除；11905 动态内容未通过安全检查；
        11906 引用活动不可用。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResult"
...
  schemas:
    Id:
      type: string
      minLength: 1
      pattern: "^[1-9][0-9]*$"
      example: "1900000000000000201"
    Instant:
      type: string
      format: date-time
      example: "2026-07-29T08:00:00Z"
    ResultMeta:
      type: object
      required: [code, message, requestId, timestamp]
      properties:
        code:
          type: integer
          enum: [0]
        message:
          type: string
          example: 成功
        requestId:
          type: string
          example: 019fa7f8b1aa7253
        timestamp:
          $ref: "#/components/schemas/Instant"
    ErrorResult:
      type: object
      required: [code, message, data, requestId, timestamp]
      properties:
        code:
          type: integer
          minimum: 1
          example: 10001
        message:
          type: string
          example: 参数校验失败
        data:
          nullable: true
          description: 参数校验失败时可能为 FieldValidationError 数组。
        requestId:
          type: string
        timestamp:
          $ref: "#/components/schemas/Instant"
...
    AgreementType:
      type: string
      enum: [USER_AGREEMENT, PRIVACY_POLICY]
    AgreementSummary:
      type: object
      required:
        - agreementId
        - type
        - versionCode
        - title
        - requiresReconsent
        - effectiveAt
        - agreed
      properties:
        agreementId:
          $ref: "#/components/schemas/Id"
        type:
          $ref: "#/components/schemas/AgreementType"
        versionCode:
          type: string
        title:
          type: string
        requiresReconsent:
          type: boolean
        effectiveAt:
          $ref: "#/components/schemas/Instant"
        agreed:
          type: boolean
          nullable: true
          description: 匿名访问时为 null。
    AgreementDetail:
      allOf:
        - $ref: "#/components/schemas/AgreementSummary"
        - type: object
          required: [content, contentHash]
          properties:
            content:
              type: string
            contentHash:
              type: string
              pattern: "^[a-fA-F0-9]{64}$"
    AgreementConsent:
      type: object
      required: [agreementId, agreedAt]
      properties:
        agreementId:
          $ref: "#/components/schemas/Id"
        agreedAt:
          $ref: "#/components/schemas/Instant"
    AgreementItemsResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/AgreementSummary"
    AgreementDetailResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/AgreementDetail"
    AgreementConsentResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/AgreementConsent"
...